requirements.txt
pandas>=2.2.0
streamlit>=1.32.0
plotly==5.17.0
openai==1.3.0
python-dotenv==1.0.0
bcrypt==4.1.1
requests==2.31.0
